Fix chargingSessions only showing one row and other layout issues
There is definitely still work to be done, but this makes the layout at
least viewable in almost all situations.
With the adding of <!DOCTYPE html> the style.height assignment in
pageNavigation stopped working due to missing “px”
The same goes for the margin on #carUpdate

Since scrollHeight does not include margins, the margin at the bottom
of chargingOverview is now being added by the observer and the 0.5em at
the top of a chargingSession element. The latter is still a bit hacky
and should be done differently.