MobileAppAPI: Fix login issue following a change by VW
The password form (credentialsForm) is unfortunately not returned in
the HTML response anymore (it is now constructed using js)
To mitigate this, we parse the data that is used to construct the form
directly out of the response (hmac and the action url for the password
form)
This code unfortunately now has to make a lot more assumptions. This
will likely cause it to break more easily with smaller changes on the
VW side.
Fixes #45